Split test to other file
